KYLIN-3995 config data type may make oom

+++ core-common/src/main/java/org/apache/kylin/common/KylinConfigBase.java
@@ -486,7 +486,7 @@ abstract public class KylinConfigBase implements Serializable {
         return Boolean.parseBoolean(getOptional("kylin.dictionary.use-forest-trie", TRUE));
     }
 
-    public int getTrieDictionaryForestMaxTrieSizeMB() {
+    public long getTrieDictionaryForestMaxTrieSizeMB() {
         return Integer.parseInt(getOptional("kylin.dictionary.forest-trie-max-mb", "500"));
     }
